==> Step 0 -Testing Measure One.ipynb <==
# Evaluating if method 1 (novelty in recombination) from Veugelers paper is feasible

Novelty in recombination: Determine if a pairwise combination of USPTO classes of a patent has occured previously, or is it the first (and thus novel) occurance

import pandas as pd

#take in data
master = pd.read_csv('USDesign.csv')
master.head()

Based off full classification

num_pats = master.patent_number.nunique()

num_pats

num_pats_multiclass = master[master.duplicated(subset=['patent_number'], keep=False)].patent_number.nunique()

num_pats_multiclass

((num_pats - num_pats_multiclass)/num_pats) * 100

~~There are 342,800 data points representing 312,050 unique patent numbers. 27,448 of these patents have more than one class associated with them at the group level, resulting in 90% of the patent population only having one class. Based on these metrics, it is **not** appropriate to use novelty in recombination as a measure.~~

~~**Important** The statement above was determined during the intial data query, which was only requesting the group level classification.~~

~~There are 358,321 data data points representing 254,113 unique patent numbers. More work will have to be done to understand why there is a discrepancy between the first and second totals of unique patent numbers. 72,758 Of these patents have classes associated with them, resulting in 71.36% of the patents only havine one class. Based on these metrics, it is not appropriate to use novelty in recombination as a measure.~~

The lastest pull represents 415,134 unique patent numbers. Of these about 72% of the patents only have one class associated with them. Based off this, novelty in recombination does not seem like the best method
