==> Step 3 - Clean Citation Data.ipynb <==
# We have the citation data. Now let's clean it up

import pandas as pd
import numpy as np
import pickle
import json

# load the object from the file into var b
cite1 = pickle.load(open('cite_files/cite1file', 'rb'))  
cite2 = pickle.load(open('cite_files/cite2file', 'rb'))
cite3 = pickle.load(open('cite_files/cite3file', 'rb'))
cite4 = pickle.load(open('cite_files/cite4file', 'rb'))
cite5 = pickle.load(open('cite_files/cite5file', 'rb'))
cite6 = pickle.load(open('cite_files/cite6file', 'rb'))
cite7 = pickle.load(open('cite_files/cite7file', 'rb'))
cite8 = pickle.load(open('cite_files/cite8file', 'rb'))
cite9 = pickle.load(open('cite_files/cite9file', 'rb'))
cite10 = pickle.load(open('cite_files/cite10file', 'rb'))
all_cites = cite1 + cite2 + cite3 + cite4 + cite5 + cite6 + cite7 + cite8 + cite9 + cite10

len(all_cites)

cite_list = []
for index, item in enumerate(all_cites):
    if (all_cites[index]['patents'] != None):
        cite_list.append(all_cites[index]['patents'][0])

len(cite_list)

output = {'patents': cite_list}

citations = pd.io.json.json_normalize(output['patents'], record_path='uspcs', meta='patent_number')
citations.head()

citations.shape

citations.groupby('patent_number').ngroups

Since we gathered this data in overlapping sections, we have some duplicate entries. Let's get rid of them

citations = citations.drop_duplicates().reset_index(drop=True)
citations.head()

citations.shape

Saving dataframe

citations.to_csv('citations.csv', index=False)
