==> Step 4 - Combining citation and class.ipynb <==
import pandas as pd
import numpy as np

import design_class_dictionary

#import relevant data
# master = pd.read_csv('data/master.csv')
master = pd.read_csv('data/master_1980_2013.csv')
citations = pd.read_csv('data/citations.csv', dtype={'patent_number':str})
focal_citation_link = pd.read_csv('data/citations_patents_level.csv')
focal_class = pd.read_csv('data/uspcs_level.csv')

master.head()

focal = master[['patent_number','priority_date']]
focal.head()

citations.head()

focal_citation_link.head()

focal_citation_link.isnull().sum()

focal_citation = pd.merge(focal, focal_citation_link, on='patent_number', how='left')
focal_citation.head()

focal_citation['patent_number'].nunique()

focal_citation.isnull().sum()

Let's see what is null and why

focal_citation.loc[focal_citation['cited_patent_number'].isnull()]

After viewing a sample, these patents either don't make references, rare, or all of their references are pre 1976, more common. Since they are not useful for analysis, I will drop them

# percentage of patents affected
focal_citation['cited_patent_number'].value_counts(normalize=True, dropna=False).head()

Less than 1 percent of patents are affected by this

focal_citation.dropna(inplace=True)
focal_citation.patent_number.nunique()

# current number of patents represened: 507,127

citations.rename(index=str, columns={'patent_number':'cited_patent_number'}, inplace=True)
focal_citation_citeclass = pd.merge(focal_citation, citations, on='cited_patent_number', how='left')
focal_citation_citeclass.patent_number.nunique()

focal_citation_citeclass.head()

focal_citation_citeclass.rename(index=str, columns={'uspc_mainclass_id':'cite_mainclass', 'uspc_subclass_id':'cite_subclass'}, inplace=True)
focal_citation_citeclass.isnull().sum()

Drop nulls. Without classification its useles

focal_citation_citeclass.dropna(inplace=True)
focal_citation_citeclass['patent_number'].nunique()

focal_citation_citeclass.loc[(focal_citation_citeclass.cite_mainclass.str.contains('D')) & (~focal_citation_citeclass.cited_patent_number.str.contains('D'))]

citations.head()

citations.loc[citations.cited_patent_number.str.match('4069641')]

# Current number of patents represented: 507,126

final = pd.merge(focal_citation_citeclass, focal_class, on='patent_number', how='left')
final.head()

final.rename(index=str, columns={'uspc_mainclass_id':'focal_mainclass', 'uspc_subclass_id':'focal_subclass'}, inplace=True)

final.shape

final.patent_number.nunique()

final.isnull().sum()

4000 of our patents had null citations. These are mainly withdrawn patents and newly granted patent. I assume this has to do with patentsview getting updated information

final.dropna(inplace=True)
final['patent_number'].nunique()

# Current patent count: 503,128

final.to_csv('data/final_no_drops.csv', index_label=False)

final = pd.read_csv('data/final_no_drops.csv')
final.head()

# final.focal_mainclass.sort_values().unique()

# final.cite_mainclass.sort_values().unique()

There are inconsitency in some of the class name. Ex: D01 and D1. This needs to be reconciled

class_dictionary = {'D1':'D01', 'D2':'D02', 'D3':'D03', 'D4':'D04', 'D5':'D05', 'D6':'D06', 'D7':'D07', 'D8':'D08', 'D9':'D09'}

# replacing mainclass level
final['cite_mainclass'] = final['cite_mainclass'].replace(class_dictionary)
final['focal_mainclass'] = final['focal_mainclass'].replace(class_dictionary)

# final.cite_mainclass.sort_values().unique()

#replacing subclass level
#easier to remove portion of string than adding in the middle contionally. D1 and D01 are equivalent anyways
final['cite_subclass'] = final['cite_subclass'].replace('D0', 'D', regex=True)
final['focal_subclass'] = final['focal_subclass'].replace('D0', 'D', regex=True)


No all patents have a typical class assignements or citations. Going to remove them from the master dataset and put them aside for further analysis

# remove cite_mainclass from no longer published, or plant

special = final.loc[(final.cite_mainclass.str.contains('No longer published')) | (final.cite_mainclass.str.contains('PLT'))]

final = final.drop(special.index)

final.patent_number.nunique()

# Current patent count: 503,117

# remove design patent with utility class and utility patent with design class
design_with_util_class = final.loc[(~final.focal_mainclass.str.contains('D'))]

final = final.drop(design_with_util_class.index)

utility_with_design_class = final.loc[(~final.cited_patent_number.str.contains('D')) & (final.cite_mainclass.str.contains('D'))]
final = final.drop(utility_with_design_class.index)

final.patent_number.nunique()

# Current patent count: 503,107

This is with very tradional design and utility classes

outcast = pd.concat([special, design_with_util_class, utility_with_design_class])

outcast.patent_number.nunique()

102,972 patents are affected by these extraordinary catogories, but only 21 patents overall were unusable

#save
# final.to_csv('data/final.csv', index_label=False)
# outcast.to_csv('data/outcast.csv', index_label=False)

# final.to_csv('data/final_481186.csv', index_label=False)
