==> Step 5 - Removing nontraditional patent classifications from master.ipynb <==
import pandas as pd
import numpy as np

master = pd.read_csv('data/master.csv')
master.head()

final = pd.read_csv('data/final.csv')
final.head()

master = master.loc[master.patent_number.isin(final.patent_number.unique().tolist())]
master.to_csv('data/master_503107.csv', index_label=False)
