==> Step 7 - Splitting data by utility vs design.ipynb <==
This purpose of this notebook is to split the data into 4 groups for analysis.
1. Design citing design at mainclass level
2. Design citing utility at mainclass level
3. Design citing deisng at subclass level
4. Design citing utility at subclass level

import pandas as pd
import numpy as np
np.set_printoptions(threshold=np.nan)

#reading in cleaned data. This csv lists each patent, their priority date, their main class, their subclass,
# the patents they cite, the cited patents mainclass, and cited patent focal class
# final =pd.read_csv('data/final.csv')
final = pd.read_csv('data/final.csv')
final.head()

master = pd.read_csv('data/master_437678.csv')

final = final.loc[final.patent_number.isin(master.patent_number.unique().tolist())]

final.patent_number.nunique()

# mainclass
First we will select just the mainclass levels

#select the mainclass aggregation
final_main = final.drop(labels = ['cite_subclass','focal_subclass'], axis=1)
final_main.head()

#select just design citing design. The cited patent number will contain a D
#check for proper classifications
final_main_d2d = final_main.loc[final_main.cited_patent_number.str.contains('D')]

print('Cited mainclasses:',np.sort(final_main_d2d.cite_mainclass.unique()),\
      '\n\n','Focal Mainclasses:', np.sort(final_main_d2d.focal_mainclass.unique()))

'''It seems as though some cited design patents have untilty classifications mixed in.
We will ignore them by only selecting citation classes that contain "D"
We will also drop the "cited patent number" column because it is no longer needed.
In order to make sure each class pair is represented once for each patent, we drop any
duplicate entries left in the dataframe.
'''
final_main_d2d = final_main_d2d.loc[final_main_d2d.cite_mainclass.str.contains('D')]\
.drop(['cited_patent_number'], axis=1)

'''
We only want to have non-repeating combinations of class pairings
So drop class pairings like A-A, B-B, C-C, etc...
Also if A-B and B-A are represented in focal patents, drop the repeating combinations
'''
#drop A-A
final_main_d2d = final_main_d2d.loc[final_main_d2d.cite_mainclass != final_main_d2d.focal_mainclass]

#drop A-B, B-A duplicates
sorted_cols = final_main_d2d[['cite_mainclass','focal_mainclass']].apply(sorted, axis=1, result_type='expand')
final_main_d2d = pd.concat([final_main_d2d, sorted_cols], axis=1)
final_main_d2d = final_main_d2d.drop_duplicates(['patent_number',0,1]).drop([0,1], axis=1)

final_main_d2d.sample(10)

'''Similar procedure as above, except for design citing utility at the mainclass
'''
#select patent numbers that do NOT comtain "D"
final_main_d2u = final_main.loc[~final_main.cited_patent_number.str.contains('D')]
print('Cited mainclass: ',np.sort(final_main_d2u.cite_mainclass.unique()), '\n\n',\
      'Cited subclass: ',np.sort(final_main_d2u.focal_mainclass.unique()))

'''All the classes seen to be correct.
As above, drop the cited patent number and any duplicates from the data frame to ensure
that each class combination is represented once for each patent
'''

#drop A-A
final_main_d2u = final_main_d2u.loc[final_main_d2u.cite_mainclass != final_main_d2u.focal_mainclass]

#drop A-B, B-A duplicates
sorted_cols = final_main_d2u[['cite_mainclass','focal_mainclass']].apply(sorted, axis=1, result_type='expand')
final_main_d2u = pd.concat([final_main_d2u, sorted_cols], axis=1)
final_main_d2u = final_main_d2u.drop_duplicates(['patent_number',0,1]).drop(['cited_patent_number',0,1], axis=1)

final_main_d2u.sample(10)

# Subclass
Repeat the same procedure as above, except at the subclass level

final_sub = final.drop(labels = ['cite_mainclass','focal_mainclass'], axis=1).drop_duplicates()
final_sub.head()

'''select just design citing design. The cited patent number will contain a D
The subclass level contains MANY more classes than the main class.
Printing them all out is slow and unecessary
'''
final_sub_d2d = final_sub.loc[final_sub.cited_patent_number.str.contains('D')]
final_sub_d2d = final_sub_d2d.loc[final_sub_d2d.cite_subclass.str.startswith('D')]\
.drop(['cited_patent_number'], axis=1)


#drop A-A
final_sub_d2d = final_sub_d2d.loc[final_sub_d2d.cite_subclass != final_sub_d2d.focal_subclass]

#drop A-B, B-A duplicates
sorted_cols = final_sub_d2d[['cite_subclass','focal_subclass']].apply(sorted, axis=1, result_type='expand')
final_sub_d2d = pd.concat([final_sub_d2d, sorted_cols], axis=1)
final_sub_d2d = final_sub_d2d.drop_duplicates(['patent_number',0,1]).drop([0,1], axis=1)

final_sub_d2d.sample(10)


#design citing utility
final_sub_d2u = final_sub.loc[~final_sub.cited_patent_number.str.contains('D')]
final_sub_d2u = final_sub_d2u.loc[~final_sub_d2u.cite_subclass.str.startswith('D')]\
.drop(['cited_patent_number'], axis=1)



#drop A-A
final_sub_d2u = final_sub_d2u.loc[final_sub_d2u.cite_subclass != final_sub_d2u.focal_subclass]

#drop A-B, B-A duplicates
sorted_cols = final_sub_d2u[['cite_subclass','focal_subclass']].apply(sorted, axis=1, result_type='expand')
final_sub_d2u = pd.concat([final_sub_d2u, sorted_cols], axis=1)
final_sub_d2u = final_sub_d2u.drop_duplicates(['patent_number',0,1]).drop([0,1], axis=1)

final_sub_d2u.sample(10)

# Stats
Quickt comparisons of the dataframes

def quick_stats(frame, focal_column, cited_column):
    count = frame.shape[0]
    unique_pats = frame.patent_number.nunique()
    focal = frame[focal_column].nunique()
    cite = frame[cited_column].nunique()
    mean_size = round(frame.groupby('patent_number').size().mean(), 2)
    
    print('Number of entries: ', count,'\n',\
         'Number of unique patents: ', unique_pats,'\n',\
         'Number of classes represented by focal patents: ',focal,'\n',\
          'Number of classes represented by cited patents: ', cite,'\n',\
          'Mean number of citation pairs per patent: ', mean_size
         )
    

print('Design from Design mainclass:')
quick_stats(final_main_d2d, 'focal_mainclass','cite_mainclass')
print(' \n Design from utility mainclass:')
quick_stats(final_main_d2u, 'focal_mainclass','cite_mainclass')
print(' \n Design from Design subclass:')
quick_stats(final_sub_d2d, 'focal_subclass','cite_subclass')
print(' \n Design from utility subclass')
quick_stats(final_sub_d2u, 'focal_subclass','cite_subclass')


Save the dataframes for future analysis


# final_main_d2d.to_csv('data/final_main_d2d.csv', index_label=False)
# final_main_d2u.to_csv('data/final_main_d2u.csv', index_label=False)


# final_sub_d2d.to_csv('data/final_sub_d2d.csv', index_label=False)
# final_sub_d2u.to_csv('data/final_sub_d2u.csv', index_label=False)

final_main_d2d.to_csv('data/final_main_d2d_drops.csv', index_label=False)
final_main_d2u.to_csv('data/final_main_d2u_drops.csv', index_label=False)
final_sub_d2d.to_csv('data/final_sub_d2d_drops.csv', index_label=False)
final_sub_d2u.to_csv('data/final_sub_d2u_drops.csv', index_label=False)
